chore: change limit type

+++ unix-utils/src/response.rs
@@ -117,7 +117,7 @@ pub struct PossibleGroup {
     pub name: String,
     pub use_inspection: bool,
     pub secondary_text: Option<String>,
-    pub limit: Option<u32>,
+    pub limit: Option<u64>,
 }
 
 #[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
